Added check to see if online.
Added a check to see if the user is online and display it on the users page. Also added a last online box on the profile page. Old users may have a broken date on their profile until they log in and it converts the updated_at timestamp for the user. Finally, added formatting fixes in the functions.php file.