Converted entire project to PDO
Converted entire project to use PDO! This will ultimately allow for cleaner code and allow it to be written much faster than mysqli. This commit also features a seen and unseen messages page as opposed to grouping them into 1 page. The next thing on our agenda is to include sent messages.