Added option to set all messages as seen
As well as that, fixed a bug in the message file that doesn't display the correct success message after sending. Also fixed a bug where messages were not being set to seen and was throwing an error every time you clicked on a message. And finally, added the error and success handling on the message page.