Empty check on status.
The website will not check if the inputted status is empty. If so, it will throw an error. Added function to purify status before sending it to database (prevents XSS) so I won't have to do it every time I display it on the website. As a final feature, I added error checking on the dashboard to display the status errors if they come up.