ip addresses + more
User IP addresses will now be collected at signup and login to give use to IP bans if the user is in extreme violation of the website's terms of service. The next thing on the agenda, is implementing IP bans themselves. Username and passwords now have limits. Password must be between 8 and 50 characters. Usernames must be between 3 and 20 characters. Fixed bug where user statuses were not displaying some characters correctly and converting them into HTML entities. The user will now be declared as offline if they are inactive for more than 3 minutes. Previously, this was 1 minute and 30 seconds.