Implemented IP bans.
If the user connecting to the site's ip address is on the IP ban list they will be redirected to literal hell. I included a very harsh ban message for the highest level of douchebaggery known to man-kind. IP bans as stated on the ban page can not be revoked.