New UI for guest users.
Changelog

New UI for guest users. They will now only be limited to the login, signup, and home pages.

Site settings (unfinished, wont go into much detail) When it is finished it will store site alerts and if the website is in maintenance or not.

User bios have been implemented, basically status but instead of a 50 character limit you get a 3000 character limit. This also supports markdown.