Timezone issue fixed.
An issue where both the last online, join date, and last online fields were being displayed incorrectly on the users and profile pages has been identified and fixed. It is unclear whether this bug also applies to the messaging feature. If it is, it will be patched out in the next release.