ip address bug fixed
Fixed bug where the hashed ip addresses in the users table were getting cut off due to the character limit on the user_ip and user_signup_ip columns. Also fixed unnecessary dehash function in check to see if ip is banned because you don't need to get anything from the users table to check if the ip is banned anyway.