.gitignore to ignore non web files
`.gitignore` now ignores the following files

`css.bat`
`css.index.bat`
`forum.sql`
`README.md`
`config\config.php`

This is so that I can download the website directly from GitHub and quickly deploy to the website.
`config\config.php` was put onto `.gitignore` due to the timezone bug which requires me to change the line that causes it immediately.

This commit also includes *the beginning of the badge feature*. Users will be allowed to gain badges as they use the site such as forum awards or friend awards.

Co-Authored-By: fanchDev <51495972+eifo1@users.noreply.github.com>