Added site maintenance & admin logs
Site Maintenance is now a feature on Brick-Town. Only admins, site developers, and the owner will be allowed onsite whilst in maintenance.

Admin actions such as updating the banner or updating the maintenance toggle will now be logged with your user id, what action you performed, and the date and time you performed it.

Other Changes:
If the website is in maintenance you will no longer be able to update the site banner.